refactor: criar pasta navbar para os componentes da navbar
