+++ src/App.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { BrowserRouter, Routes, Route, Link } from 'react-router-dom'
 import './App.css';
-import NavBar from './components/navBar'
+import NavBar from './components/navbar/navBar';
 import Home from './pages/Home'
 import Table from './pages/Table'
